+++ changelog.md
@@ -1,3 +1,7 @@
-PIPEDRIVE [handlers] (2020-05-07) 
+Pipedrive [handlers] (2020-05-07) 
   [pipedrive_api_v1] (2020-05-07)
     * Initial commit of the handler
+
+Pipedrive [bridge-adapters] (2020-06-04)
+  [kinetic-bridgehub-adapter-pipedrive] (2020-06-04)
+    * Initial commit of adapter